+++ src/account.h
@@ -14,7 +14,7 @@ struct game_account
 
 /* creation */
 ACCOUNT_DATA	*init_account( void );
-ACCOUNT_DATA 	*load_accountByID( int id );
+ACCOUNT_DATA 	*load_accountByID( long long id );
 ACCOUNT_DATA	*load_accountByName( const char *name );
 int		 db_load_account( ACCOUNT_DATA *account, MYSQL_ROW *row );
 int		 new_account( ACCOUNT_DATA *account );
@@ -30,8 +30,8 @@ int		 set_aPasswd( ACCOUNT_DATA *account, const char *oldpw, const char *passwd 
 int		 set_aLevel( ACCOUNT_DATA *account, ACCT_LEVEL level );
 
 /* getters */
-ACCOUNT_DATA	*get_accountByID( int id ); /* load and attach if not already active */
-ACCOUNT_DATA	*get_accountByID_ifActive( int id );
+ACCOUNT_DATA	*get_accountByID( long long id ); /* load and attach if not already active */
+ACCOUNT_DATA	*get_accountByID_ifActive( long long id );
 ACCOUNT_DATA	*get_accountByName( const char *name ); /* ibid */
 ACCOUNT_DATA	*get_accountByName_ifActive( const char *name );
 D_SOCKET	*get_aSocket( ACCOUNT_DATA *account );
